[chore]: Move `run_async` to `_internal/utils`
Part of an upcoming refactoring, moved to a
separate PR for a smaller diff and to avoid merge
conflicts. `run_async` will be used both in
`_internal/server` and `_internal/proxy`.
