+++ src/dstack/_internal/utils/common.py
@@ -1,11 +1,23 @@
+import asyncio
 import itertools
 import re
 import time
 from collections.abc import Callable
 from datetime import datetime, timedelta, timezone
+from functools import partial
 from pathlib import Path
 from typing import Any, Iterable, List, Optional, TypeVar
 from urllib.parse import urlparse
+
+from typing_extensions import ParamSpec
+
+P = ParamSpec("P")
+R = TypeVar("R")
+
+
+async def run_async(func: Callable[P, R], *args: P.args, **kwargs: P.kwargs) -> R:
+    func_with_args = partial(func, *args, **kwargs)
+    return await asyncio.get_running_loop().run_in_executor(None, func_with_args)
 
 
 def get_dstack_dir() -> Path:
